Create Genesis. Restored.
This is the start of a Restoration process. Bringing the Books of YHWH back to its roots. This is Genesis. Fully refined back towards its original context before the empire system tried to generically his name as well as change the name of Yeshua HaMashiach into the idol we know today as Jesus. Which is in fact a incorrect translation and a deceptive one at that. Read, and study for yourself.